+++ jalgs/besselI.java
@@ -18,8 +18,8 @@ public class besselI{
 		bessel1vals=new double[10000];
 		bessel2vals=new double[10000];
 		for(int i=0;i<10000;i++){
-			bessel0vals[i]=bessi0(0.1*(double)i);
-			bessel1vals[i]=bessi1(0.1*(double)i);
+			bessel0vals[i]=bessi0(0.1*i);
+			bessel1vals[i]=bessi1(0.1*i);
 			// bessel2vals[i]=bessel2(0.1*(double)i);
 		}
 	}
@@ -46,7 +46,7 @@ public class besselI{
 				nextval=bessel2vals[index+1];
 			}
 		}
-		fraction=dindex-(double)index;
+		fraction=dindex-index;
 		return(prevval+fraction*(nextval-prevval));
 	}
